% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{somar duas parcelas x e y}
\usage{
somar(x, y)
}
\arguments{
\item{x}{parcela x}

\item{y}{parcela y}
}
\value{
retorna soma das parcelas x e y
}
\description{
somar duas parcelas x e y
}
\examples{

somar(3,5)
}
